Pythagorean tree: Move util classes to util folder and remove widgetutils
