+++ Orange/widgets/visualize/utils/owgrid.py
