+++ Orange/widgets/visualize/utils/owlegend.py
