+++ Orange/widgets/visualize/utils/scene.py
