+++ Orange/widgets/visualize/utils/tree/rules.py
