+++ Orange/widgets/visualize/utils/tree/treeadapter.py
