+++ Orange/widgets/visualize/utils/view.py
